+++ src/eos.rs
@@ -9,12 +9,11 @@ use crate::saftvrqmie::SaftVRQMie;
 #[cfg(feature = "uvtheory")]
 use crate::uvtheory::UVTheory;
 use feos_core::cubic::PengRobinson;
-use feos_core::joback::Joback;
 #[cfg(feature = "python")]
-use feos_core::python::user_defined::{PyIdealGas, PyResidual};
+use feos_core::python::user_defined::PyResidual;
 use feos_core::si::*;
 use feos_core::*;
-use feos_derive::{Components, IdealGas, Residual};
+use feos_derive::{Components, Residual};
 use ndarray::Array1;
 
 /// Collection of different [Residual] implementations.
@@ -39,16 +38,4 @@ pub enum ResidualModel {
     Pets(Pets),
     #[cfg(feature = "uvtheory")]
     UVTheory(UVTheory),
-}
-
-/// Collection of different [IdealGas] implementations.
-///
-/// Particularly relevant for situations in which generic types
-/// are undesirable (e.g. FFI).
-#[derive(Components, IdealGas)]
-pub enum IdealGasModel {
-    NoModel(usize),
-    Joback(Joback),
-    #[cfg(feature = "python")]
-    Python(PyIdealGas),
 }