+++ src/python/eos.rs
@@ -1,10 +1,11 @@
-use crate::eos::{IdealGasModel, ResidualModel};
+use crate::eos::ResidualModel;
 #[cfg(feature = "estimator")]
 use crate::estimator::*;
 #[cfg(feature = "gc_pcsaft")]
 use crate::gc_pcsaft::python::PyGcPcSaftEosParameters;
 #[cfg(feature = "gc_pcsaft")]
 use crate::gc_pcsaft::{GcPcSaft, GcPcSaftOptions};
+use crate::ideal_gas::IdealGasModel;
 #[cfg(feature = "estimator")]
 use crate::impl_estimator;
 #[cfg(all(feature = "estimator", feature = "pcsaft"))]
@@ -26,10 +27,10 @@ use crate::uvtheory::python::PyUVParameters;
 #[cfg(feature = "uvtheory")]
 use crate::uvtheory::{Perturbation, UVTheory, UVTheoryOptions, VirialOrder};
 
+use super::dippr::PyDippr;
+use super::joback::PyJoback;
 use feos_core::cubic::PengRobinson;
-use feos_core::joback::Joback;
 use feos_core::python::cubic::PyPengRobinsonParameters;
-use feos_core::python::joback::PyJobackParameters;
 use feos_core::python::user_defined::{PyIdealGas, PyResidual};
 use feos_core::si::*;
 use feos_core::*;
@@ -316,14 +317,29 @@ impl PyEquationOfState {
     ///
     /// Parameters
     /// ----------
-    /// parameters : List[JobackRecord]
-    ///     List containing
+    /// joback : Joback
+    ///     The parametrized Joback model.
     ///
     /// Returns
     /// -------
     /// EquationOfState
-    fn joback(&self, parameters: PyJobackParameters) -> Self {
-        self.add_ideal_gas(IdealGasModel::Joback(Joback::new(parameters.0)))
+    fn joback(&self, joback: PyJoback) -> Self {
+        self.add_ideal_gas(IdealGasModel::Joback(joback.0))
+    }
+
+    /// Ideal gas model based on DIPPR equations for the ideal
+    /// gas heat capacity.
+    ///
+    /// Parameters
+    /// ----------
+    /// dippr : Dippr
+    ///     The parametrized Dippr model.
+    ///
+    /// Returns
+    /// -------
+    /// EquationOfState
+    fn dippr(&self, dippr: PyDippr) -> Self {
+        self.add_ideal_gas(IdealGasModel::Dippr(dippr.0))
     }
 }